Optimized user search api

+++ src/main/java/top/xsd666/usercenterbackend/controller/UserController.java
@@ -9,7 +9,10 @@ import top.xsd666.usercenterbackend.model.User;
 import top.xsd666.usercenterbackend.model.request.UserBasicRequest;
 import top.xsd666.usercenterbackend.service.UserService;
 
+import java.util.ArrayList;
 import java.util.List;
+
+import static top.xsd666.usercenterbackend.constant.UserConstant.*;
 
 @RestController
 @RequestMapping("/user")
@@ -54,19 +57,33 @@ public class UserController {
     }
 
     @GetMapping("/{phone}")
-    public List<User> getUsers(@PathVariable String phone) {
+    public List<User> getUsers(@PathVariable String phone, HttpServletRequest request) {
+        if (!this.isAdmin(request)) {
+            return new ArrayList<>();
+        }
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(phone)) {
             queryWrapper.eq("phone", phone);
         }
-        return this.userService.list(queryWrapper);
+        List<User> users = this.userService.list(queryWrapper);
+        users.forEach(this.userService::desensitize);
+        return users;
     }
 
     @DeleteMapping("/{id}")
-    public Boolean deleteUser(@PathVariable long id) {
+    public Boolean deleteUser(@PathVariable long id, HttpServletRequest request) {
+        if (!this.isAdmin(request)) {
+            return false;
+        }
         if (id <= 0) {
             return false;
         }
         return this.userService.removeById(id);
+    }
+
+
+    private boolean isAdmin(HttpServletRequest request) {
+        User user = (User) request.getSession().getAttribute(USER_LOGIN_STATE);
+        return user != null && user.getUserRole() == ADMIN_ROLE;
     }
 }
